make tx signing work

+++ src/core/definitions/mod.rs
@@ -146,10 +146,10 @@ impl ToBytes for FieldHeader {
 
         if self.type_code < 16 {
             if self.field_code < 16 {
-                let shift = self.type_code << 4 | self.field_code;
+                let shift = (self.type_code << 4 | self.field_code) as u8;
                 header_bytes.extend_from_slice(&shift.to_be_bytes());
             } else {
-                let shift = self.type_code << 4;
+                let shift = (self.type_code << 4) as u8;
 
                 header_bytes.extend_from_slice(&shift.to_be_bytes());
                 header_bytes.extend_from_slice(&self.field_code.to_be_bytes());
